add a version number to calls to nike API
It seems that calls to nike API needs a version number before between the domain name and the details of the path. For example:
https://api.nike.com/v1/me/sport/activities

On a call like "https://api.nike.com/me/sport/activities" I get a 401 "Unauthorized" with a body of:
<error>invalid_token</error>

adding /v1 of to the path seems to do the trick.